fix: Abstracted drag and drop widget
